move configmaps
